# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AUIB Academic Advisor API",
    "description": "Degree planning for AUIB students. Planning endpoints are stateless and store nothing.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/history/parse": {
      "post": {
        "tags": [
          "planner"
        ],
        "summary": "Read a pasted SIS Course History page (F11.3)",
        "operationId": "parse_history_api_v1_history_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoryParseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryParseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/planner/progress": {
      "post": {
        "tags": [
          "planner"
        ],
        "summary": "Progress per requirement and what is left (F1.2, F4)",
        "operationId": "progress_api_v1_planner_progress_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/planner/plan": {
      "post": {
        "tags": [
          "planner"
        ],
        "summary": "A term-by-term plan to graduation (F1.3, F1.4, F1.6)",
        "operationId": "plan_api_v1_planner_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/planner/what-if": {
      "post": {
        "tags": [
          "planner"
        ],
        "summary": "What dropping or delaying a course does to graduation (F1.5)",
        "operationId": "plan_what_if_api_v1_planner_what_if_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatIfOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/planner/recommendations": {
      "post": {
        "tags": [
          "planner"
        ],
        "summary": "Ranked electives for each open requirement (F2.2)",
        "operationId": "recommendations_api_v1_planner_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/planner/gpa": {
      "post": {
        "tags": [
          "planner"
        ],
        "summary": "Project the CGPA from expected grades, or the grades a target CGPA needs (F7.2, F7.3)",
        "operationId": "gpa_plan_api_v1_planner_gpa_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GpaPlanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GpaPlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseRef": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "units": {
            "type": "number",
            "title": "Units"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "units"
        ],
        "title": "CourseRef"
      },
...
      "GpaPlanIn": {
        "properties": {
          "program_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Program Id"
          },
          "minor_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Minor Id"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/AttemptIn"
            },
            "type": "array",
            "maxItems": 300,
            "title": "Attempts"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesIn"
          },
          "entry_term": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40,
                "minLength": 6,
                "examples": [
                  "Fall 2027",
                  "2027/2028 Fall"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Term",
            "description": "The term the student joined AUIB; worked out from the course history if empty"
          },
          "program_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Program Version",
            "description": "A specific version of the major, when the registrar approved a move to it (F0.4)"
          },
          "courses": {
            "items": {
              "$ref": "#/components/schemas/ExpectedGradeIn"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Courses"
          },
          "target": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 4.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "program_id"
        ],
        "title": "GpaPlanIn",
        "description": "Courses to project (F7.2) and, optionally, a target CGPA to work back from (F7.3)."
      },
      "GpaPlanOut": {
        "properties": {
          "current": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "projected": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected",
            "description": "The CGPA with the expected grades; null when none is given"
          },
          "courses_gpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Courses Gpa",
            "description": "The GPA of the courses with an expected grade alone"
          },
          "graded_credits": {
            "type": "number",
            "title": "Graded Credits"
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GpaTargetOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          }
        },
        "type": "object",
        "required": [
          "current",
          "projected",
          "courses_gpa",
          "graded_credits",
          "target",
          "assumptions"
        ],
        "title": "GpaPlanOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryParseIn": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 300000,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "HistoryParseIn"
      },
      "HistoryParseOut": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/HistoryRowOut"
            },
            "type": "array",
            "title": "Rows"
          },
          "unread": {
            "items": {
              "$ref": "#/components/schemas/UnreadLineOut"
            },
            "type": "array",
            "title": "Unread"
          },
          "ignored_line_count": {
            "type": "integer",
            "title": "Ignored Line Count"
          },
          "duplicates_removed": {
            "type": "integer",
            "title": "Duplicates Removed"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "unread",
          "ignored_line_count",
          "duplicates_removed"
        ],
        "title": "HistoryParseOut"
      },
      "HistoryRowOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "term": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "units": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Units"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AttemptStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "sis_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sis Status"
          },
          "line": {
            "type": "integer",
            "title": "Line"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "term",
          "grade",
          "units",
          "status",
          "sis_status",
          "line",
          "issues"
        ],
        "title": "HistoryRowOut"
      },
...
      "PlanOut": {
        "properties": {
          "start_term": {
            "$ref": "#/components/schemas/TermOut"
          },
          "graduation_term": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TermOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "on_time_term": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TermOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "terms": {
            "items": {
              "$ref": "#/components/schemas/PlannedTermOut"
            },
            "type": "array",
            "title": "Terms"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueOut"
            },
            "type": "array",
            "title": "Issues"
          },
          "unscheduled": {
            "items": {
              "$ref": "#/components/schemas/PlanItemOut"
            },
            "type": "array",
            "title": "Unscheduled"
          },
          "critical_chain": {
            "items": {
              "$ref": "#/components/schemas/CourseRef"
            },
            "type": "array",
            "title": "Critical Chain"
          },
          "progress": {
            "$ref": "#/components/schemas/ProgressOut"
          },
          "progress_with_plan": {
            "$ref": "#/components/schemas/GroupProgressOut"
          },
          "eligible_next_term": {
            "items": {
              "$ref": "#/components/schemas/EligibleOut"
            },
            "type": "array",
            "title": "Eligible Next Term"
          },
          "degree_map": {
            "$ref": "#/components/schemas/DegreeMapOut"
          },
          "gpa": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GpaOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null until the student has a graded course"
          },
          "minor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MinorPlanOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when no minor was chosen"
          },
          "catalog": {
            "$ref": "#/components/schemas/CatalogInfoOut"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "start_term",
          "graduation_term",
          "on_time_term",
          "terms",
          "issues",
          "unscheduled",
          "critical_chain",
          "progress",
          "progress_with_plan",
          "eligible_next_term",
          "degree_map",
          "gpa",
          "catalog",
          "assumptions",
          "disclaimer"
        ],
        "title": "PlanOut"
      },
...
      "ProgressOut": {
        "properties": {
          "percent_complete": {
            "type": "number",
            "title": "Percent Complete"
          },
          "completed_units": {
            "type": "number",
            "title": "Completed Units"
          },
          "in_progress_units": {
            "type": "number",
            "title": "In Progress Units"
          },
          "root": {
            "$ref": "#/components/schemas/GroupProgressOut"
          },
          "not_counted": {
            "items": {
              "$ref": "#/components/schemas/CourseRef"
            },
            "type": "array",
            "title": "Not Counted"
          },
          "whats_left": {
            "items": {
              "$ref": "#/components/schemas/LeftItemOut"
            },
            "type": "array",
            "title": "Whats Left"
          }
        },
        "type": "object",
        "required": [
          "percent_complete",
          "completed_units",
          "in_progress_units",
          "root",
          "not_counted",
          "whats_left"
        ],
        "title": "ProgressOut"
      },
...
      "RecommendationsOut": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupSuggestionsOut"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "type": "object",
        "required": [
          "groups"
        ],
        "title": "RecommendationsOut"
      },
...
      "StudentIn": {
        "properties": {
          "program_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Program Id"
          },
          "minor_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Minor Id"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/AttemptIn"
            },
            "type": "array",
            "maxItems": 300,
            "title": "Attempts"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesIn"
          },
          "entry_term": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40,
                "minLength": 6,
                "examples": [
                  "Fall 2027",
                  "2027/2028 Fall"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Term",
            "description": "The term the student joined AUIB; worked out from the course history if empty"
          },
          "program_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Program Version",
            "description": "A specific version of the major, when the registrar approved a move to it (F0.4)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "program_id"
        ],
        "title": "StudentIn",
        "description": "Everything the planner needs. Nothing here is stored (F11.5)."
      },
...
      "UnreadLineOut": {
        "properties": {
          "line": {
            "type": "integer",
            "title": "Line"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "line",
          "text",
          "reason"
        ],
        "title": "UnreadLineOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WhatIfIn": {
        "properties": {
          "program_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Program Id"
          },
          "minor_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Minor Id"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/AttemptIn"
            },
            "type": "array",
            "maxItems": 300,
            "title": "Attempts"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesIn"
          },
          "entry_term": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40,
                "minLength": 6,
                "examples": [
                  "Fall 2027",
                  "2027/2028 Fall"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Term",
            "description": "The term the student joined AUIB; worked out from the course history if empty"
          },
          "program_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Program Version",
            "description": "A specific version of the major, when the registrar approved a move to it (F0.4)"
          },
          "change": {
            "$ref": "#/components/schemas/ChangeIn"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "program_id",
          "change"
        ],
        "title": "WhatIfIn"
      },
      "WhatIfOut": {
        "properties": {
          "change": {
            "$ref": "#/components/schemas/ChangeIn"
          },
          "terms_later": {
            "type": "integer",
            "title": "Terms Later"
          },
          "before_graduation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TermOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "after_graduation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TermOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "shifts": {
            "items": {
              "$ref": "#/components/schemas/ShiftOut"
            },
            "type": "array",
            "title": "Shifts"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanOut"
          }
        },
        "type": "object",
        "required": [
          "change",
          "terms_later",
          "before_graduation",
          "after_graduation",
          "shifts",
          "plan"
        ],
        "title": "WhatIfOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Admin token (ADVISOR_ADMIN_TOKEN)",
        "scheme": "bearer"
      }
    }
  }
}